set up architecutre of Auth
