+++ app/src/main/java/io/zluan/ghub/network/auth/AuthService.kt
@@ -1,0 +1,5 @@
+package io.zluan.ghub.network.auth
+
+interface AuthService {
+// TODO(zluan): looks like there are 2 ways to authenticate, plain username/password vs. OAuth
+}
