+++ app/src/main/java/io/zluan/ghub/persistence/AccountDao.kt
@@ -1,0 +1,4 @@
+package io.zluan.ghub.persistence
+
+interface AccountDao {
+}
