+++ app/src/main/java/io/zluan/ghub/persistence/AppDatabase.kt
@@ -1,0 +1,5 @@
+package io.zluan.ghub.persistence
+
+abstract class AppDatabase {
+
+}
