+++ app/src/main/java/io/zluan/ghub/repository/auth/AuthRepository.kt
@@ -1,0 +1,4 @@
+package io.zluan.ghub.repository.auth
+
+class AuthRepository {
+}
