+++ app/src/main/java/io/zluan/ghub/ui/auth/AuthActivity.kt
@@ -1,0 +1,6 @@
+package io.zluan.ghub.ui.auth
+
+import androidx.appcompat.app.AppCompatActivity
+
+class AuthActivity : AppCompatActivity() {
+}
