+++ app/src/main/java/io/zluan/ghub/ui/auth/AuthViewModel.kt
@@ -1,0 +1,4 @@
+package io.zluan.ghub.ui.auth
+
+class AuthViewModel {
+}
