+++ app/src/main/java/io/zluan/ghub/ui/main/about/AboutFragment.kt
@@ -1,4 +1,4 @@
-package io.zluan.ghub.ui.settings
+package io.zluan.ghub.ui.main.about
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,19 +10,19 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import io.zluan.ghub.R
 
-class SettingsFragment : Fragment() {
+class AboutFragment : Fragment() {
 
-    private lateinit var settingsViewModel: SettingsViewModel
+    private lateinit var aboutViewModel: AboutViewModel
 
     override fun onCreateView(
             inflater: LayoutInflater,
             container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
-        settingsViewModel = ViewModelProvider(this).get(SettingsViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_slideshow, container, false)
-        val textView: TextView = root.findViewById(R.id.text_slideshow)
-        settingsViewModel.text.observe(viewLifecycleOwner, Observer {
+        aboutViewModel = ViewModelProvider(this).get(AboutViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
+        val textView: TextView = root.findViewById(R.id.text_gallery)
+        aboutViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
