+++ app/src/main/java/io/zluan/ghub/ui/main/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package io.zluan.ghub.ui.settings
+package io.zluan.ghub.ui.main.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SettingsViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
