+++ app/src/main/java/io/zluan/ghub/ui/main/settings/SettingsViewModel.kt
@@ -1,13 +1,13 @@
-package io.zluan.ghub.ui.home
+package io.zluan.ghub.ui.main.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class SettingsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
