+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="io.zluan.ghub.ui.home.HomeFragment"
+        android:name="io.zluan.ghub.ui.main.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_about"
-        android:name="io.zluan.ghub.ui.about.AboutFragment"
+        android:name="io.zluan.ghub.ui.main.about.AboutFragment"
         android:label="@string/menu_about"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_settings"
-        android:name="io.zluan.ghub.ui.settings.SettingsFragment"
+        android:name="io.zluan.ghub.ui.main.settings.SettingsFragment"
         android:label="@string/menu_settings"
         tools:layout="@layout/fragment_slideshow" />
 </navigation>
